+++ internal/dialog_message.go
@@ -22,6 +22,6 @@ type (
 		RelativeQuestions []string `json:"relative_questions"`
 		DatabaseFile      string   `json:"database_file"`
 		DatabaseFilePart  string   `json:"database_file_part"`
-		Confidence        int      `json:"confidence"`
+		Confidence        float64  `json:"confidence"`
 	}
 )
